Execute Git directly after having obtained a clean shell environment
Extracted shell and environment logic into get-shell and get-shell-env modules. Updated hooks-proxy and with-hooks-env to use the new shellEnv, improving reliability and separation of concerns. Added support for passing shell environment to hook proxy and locating git executable using the correct PATH.
