+++ app/src/lib/hooks/hooks-proxy.ts
@@ -1,52 +1,16 @@
 import { spawn } from 'child_process'
 import { randomBytes } from 'crypto'
 import { createWriteStream } from 'fs'
-import memoizeOne from 'memoize-one'
 import { basename, join } from 'path'
 import { ProcessProxyConnection } from 'process-proxy'
-import { Shescape } from 'shescape'
 import { Writable } from 'stream'
 import { pipeline } from 'stream/promises'
-import { execFile } from '../exec-file'
 
 const hooksUsingStdin = ['post-rewrite']
 
 const debug = (message: string, error?: Error) => {
   log.debug(`hooks: ${message}`, error)
 }
-
-const getShell = () => {
-  // TODO: Windows:
-  if (__WIN32__) {
-    throw new Error('Not implemented')
-  }
-
-  if (process.env.SHELL) {
-    try {
-      return {
-        shell: process.env.SHELL,
-        args: ['-ilc'],
-        ...getQuoteFn(process.env.SHELL),
-      }
-    } catch (err) {
-      debug('Failed resolving shell', err)
-    }
-  }
-
-  return {
-    shell: '/bin/sh',
-    args: ['-ilc'],
-    ...getQuoteFn('/bin/sh'),
-  }
-}
-
-const getQuoteFn = memoizeOne((shell: string) => {
-  const shescape = new Shescape({ shell, flagProtection: false })
-  return {
-    escape: shescape.escape.bind(shescape),
-    quote: shescape.quote.bind(shescape),
-  }
-})
 
 const waitForWritableFinished = (stream: Writable) => {
   return new Promise<void>(resolve => {
@@ -77,35 +41,12 @@ const exitWithError = (
   })
 }
 
-const getCleanShellEnv = async (): Promise<Record<string, string>> => {
-  const ext = __WIN32__ ? '.exe' : ''
-  const printenvzPath = join(__dirname, `printenvz${ext}`)
-
-  const { shell, args, quote } = getShell()
-  const { stdout } = await execFile(shell, [...args, quote(printenvzPath)], {
-    env: {},
-  })
-
-  return Object.fromEntries(
-    stdout.split('\0').map(line => {
-      const eqIndex = line.indexOf('=')
-      if (eqIndex === -1) {
-        throw new Error(`Invalid env var line: ${line}`)
-      }
-      const key = line.substring(0, eqIndex)
-      const value = line.substring(eqIndex + 1)
-      return [key, value]
-    })
-  )
-}
-
 export const createHooksProxy = (
   repoHooks: string[],
   tmpDir: string,
-  gitPath: string
+  gitPath: string,
+  shellEnv: Record<string, string | undefined>
 ) => {
-  const cleanShellEnv = memoizeOne(getCleanShellEnv)
-
   return async (connection: ProcessProxyConnection) => {
     const startTime = Date.now()
     const proxyArgs = await connection.getArgs()
@@ -127,6 +68,10 @@ export const createHooksProxy = (
       // user-configured but since we're executing the hook in a separate
       // shell with login it would just get re-initialized there anyway.
       'GIT_CONFIG_PARAMETERS',
+
+      'GIT_ASKPASS',
+      'GIT_SSH_COMMAND',
+      'GIT_USER_AGENT',
     ])
 
     const safeEnv = Object.fromEntries(
@@ -166,8 +111,6 @@ export const createHooksProxy = (
       '--',
       ...proxyArgs.slice(1),
     ]
-    const shellEnv = await cleanShellEnv()
-
     const { code } = await new Promise<{
       code: number | null
       signal: NodeJS.Signals | null
@@ -181,7 +124,10 @@ export const createHooksProxy = (
         signal: abortController.signal,
       })
         .on('close', (code, signal) => resolve({ code, signal }))
-        .on('error', reject)
+        .on('error', err => {
+          debug(`failed to spawn hook process:`, err)
+          reject(err)
+        })
 
       // hooks never write to stdout
       // https://github.com/git/git/blob/4cf919bd7b946477798af5414a371b23fd68bf93/hook.c#L73C6-L73C22
