+++ app/src/lib/hooks/with-hooks-env.ts
@@ -8,6 +8,7 @@ import { enableHooksEnvironment } from '../feature-flag'
 import type { IGitExecutionOptions } from '../git/core'
 import { getRepoHooks } from './get-repo-hooks'
 import { createHooksProxy } from './hooks-proxy'
+import { getShellEnv } from './get-shell-env'
 
 export async function withHooksEnv<T>(
   fn: (env: Record<string, string | undefined> | undefined) => Promise<T>,
@@ -31,22 +32,27 @@ export async function withHooksEnv<T>(
     return fn(options?.env)
   }
 
-  const gitPath = await which('git', { nothrow: true })
+  const shellEnv = await getShellEnv()
 
-  if (gitPath === null) {
-    log.info(`Git executable not found in PATH, skipping hook interception.`)
-    return fn(options?.env)
-  }
+  // TODO: will throw
+  const gitPath = await which('git', {
+    path: shellEnv.PATH,
+    pathExt: shellEnv.PATHEXT,
+  })
 
   const ext = __WIN32__ ? '.exe' : ''
   const processProxyPath = join(__dirname, `process-proxy${ext}`)
 
   const token = crypto.randomUUID()
   const tmpHooksDir = await mkdtemp(join(tmpdir(), 'desktop-git-hooks-'))
-  const hooksProxy = createHooksProxy(repoHooks, tmpHooksDir, gitPath)
+  const hooksProxy = createHooksProxy(repoHooks, tmpHooksDir, gitPath, shellEnv)
 
   const server = createProxyProcessServer(
-    conn => hooksProxy(conn).catch(() => conn.exit(1).catch(() => {})),
+    conn =>
+      hooksProxy(conn).catch(err => {
+        log.error(`hooks proxy failed:`, err)
+        conn.exit(1).catch(() => {})
+      }),
     { validateConnection: async receivedToken => receivedToken === token }
   )
   const port = await new Promise<number>((resolve, reject) => {
